Divide visitors to different files
